Factor out `once_per` parts to `utils`
